Fix minor issues

+++ nf/dmz/dmz_config.c
@@ -180,8 +180,8 @@ void dmz_print_config(struct dmz_config* config)
 	free(intra_addr_str);
 	free(intra_mask_str);
 
-	uint8_t nb_devices = rte_eth_dev_count();
-	for (uint8_t dev = 0; dev < nb_devices; dev++) {
+	uint16_t nb_devices = rte_eth_dev_count();
+	for (uint16_t dev = 0; dev < nb_devices; dev++) {
 		char* dev_mac_str = nf_mac_to_str(&(config->device_macs[dev]));
 		char* end_mac_str = nf_mac_to_str(&(config->endpoint_macs[dev]));
